Replaces lantern F-key with calling OnFKeyPressed
The pressing the F-key which previously activated/deactivated the lantern, now creates a zero second timer called OnFKeyPressed_timer which calls a function called OnFKeyPressed within engine scripts. This allows for the level programmer to have a status update event or any number of events to be called when the F key is pressed within amnesia The Dark Descent custom stories. Syntax within engine scripts is as follows:

void OnFKeyPressed(string &in asTimer)
{
// your code here.
}